==> R/00_doit.R <==
# Run all scripts ---------------------------------------------------------
source(file = "R/01_load.R")
source(file = "R/02_clean.R")
source(file = "R/03_augment.R")
source(file = "R/04_analysis_i_visualization.R")
source(file = "R/05_analysis_ii_logistic.R")
source(file = "R/06_analysis_iii_PCA.R")
source(file = "R/07_analysis_iv_Kmeans.R")
source(file = "R/08_summary.R")
rmarkdown::render(input = "doc/Presentation.Rmd")
